+++ tg_msg_manager/application/__init__.py
@@ -1,0 +1,13 @@
+"""Application runtime assembly helpers."""
+
+from .resources import RuntimeResourceFactory
+from .session import ApplicationSession, ApplicationSessionLockError
+from .services import ServiceBundle, create_service_bundle
+
+__all__ = [
+    "ApplicationSession",
+    "ApplicationSessionLockError",
+    "RuntimeResourceFactory",
+    "ServiceBundle",
+    "create_service_bundle",
+]